refactor(redis_client): extract RedisConnection from RedisClient
Pure refactor, no behavior change. Pulls socket lifecycle (open, auth,
reconnect-on-drop, close) out of RedisClient and into a new
RedisConnection class that lives in the same library via `part`.

RedisClient continues to own command-level policy (retry, timeout) and
delegates transport to the connection. Public API surface and observable
behavior are unchanged; existing tests pass against a live Redis without
modification.

This sets the seam for a follow-up that introduces opt-in connection
pooling and a `withConnection` API for transactions and pipelining.
