+++ packages/redis_client/pubspec.yaml
@@ -1,6 +1,6 @@
 name: shorebird_redis_client
 description: A lightweight Dart client library for communicating with a Redis server. Built by Shorebird.
-version: 0.0.12
+version: 0.0.13
 homepage: https://shorebird.dev
 repository: https://github.com/shorebirdtech/shorebird/tree/main/packages/redis_client
 topics: [redis, cache, shorebird]
